remove or change lib.plugin related test
